Create project directory structure & rename 'voronoi.py' to '__init__.py'
A proper directory structure is created for the project. 'voronoi.py' file
was renamed to '__init__.py'. Now the visualization files are saved in
the curent working directory within new 'visualization' directory
containing directories based on name of material.
Each test case is saved as a new file and all the test cases are saved
in the directory 'test_cases'. Also each function is saved as a new file and
is imported into '__init__.py' file of the src directory.