[ONLY FOR 2D] Use Scipy Voronoi function to create tessellations
Scipy Voronoi function was used to create tessellations. Currently it
can be used only for 2D case. The structural characteristics module was
accordingly modified such that junction lengths and distance between
neighbors could be computed for Scipy Voronoi function. The user defined
custom cost function was also developed accordingly. This shows that the
code is modular and modules can be replaced by the user as per his
needs.