Make main_run() and test_run() independent of Click
The functions main() and test() are now isolated such that it calls
main_run() and test_run() respectively with all the Click inputs. This
is done so that the functions main_run() and test_run() which perform
actual tasks can be imported as normal functions in any other external
python scripts.