Move all modules to a subdirectory 'ppp2019_optimizedmicrostructuregeneration'
For the package to work, it was important to execute the tool while
being outside of the package directory. Hence a new subdirectory with the
name 'ppp2019_optimizedmicrostructuregeneration' was created and all
modules were moved into it. Also the 'execute' bash script was edited
accordingly. The subdirectory name would be changed to our final package
name at some later stage and accordingly the import statements will also
be modified.