Generate/assign orientation to grains in main.py
It was observed that when optimization of textural characteristics was
performed, the orientations of grains were generated during each iteration.
This was a major bug due to which optimization of seed coordinates had
no effect on optimization of textural characteristics. This issue is
solved by generating orientations for grains in main.py at once and then
optimizing seed coordinates.