Use blit function of matplotlib for fast live plotting
When profiling was done, it was observed that live plotting required
consumed major chunk of execution time. It was  due to the fact that
during each iteration entire plot was updated including axes, line etc.
The code is now updated to use Blit function of matplotlib. Due to this
only the line plot is updated during each iteration and the axes are
extended further in intervals. This has resulted in very high speedup
also there is not much slowdown at higher iterations.