Add option '--save_interval' & save seeds data with extension of axes of live plot
A new Click option 'save_interval' is added. It is the frequency of
iterations of optimization, after which seeds data is stored for that
interval of iteration. The axes of the live plot is also updated at this
frequency with this integer interval value along X axis. The remaining
iterations seeds data if any, is written to file after optimization is
completed.