Update the element of args list instead of appending
A major bug was identified due to which entire RAM was getting used up
when high number of optimization iterations were performed. It was
identified that appending to args list was the cause and hence list
element is now being replaced instead of appending. This maintains the
size of list and accordingly the memory usage.