Check if data for grain-neighbor pair already exists
The functionality that computed disorientation angles and CSL type was
found to computing same data twice. In order to reduce this effort,
it is now check if the data for grain-neighbor pair exists and if they
do then that pair is skipped. This should reduce the computation effort
and time drastically.