Ensure seeds are placed at central Z plane in 2D case
The seeds were placed on upper Z plane of the simulation box i.e., all
seeds were placed at length of simulation box along Z axis. This was
modified such that the seeds are placed at half the length of simulation
box along Z axis. This change is important because the tessellations are
expected to happen along both directions of Z axis.